refactor: break  implementation into a module
