+++ src/commands/help.rs
@@ -5,11 +5,13 @@
 //! The module is structured to dispatch the help command based on user-provided topics,
 //! and it includes functions for generating formatted, colored output for the terminal.
 
+use crate::help_details::edit_help::show_edit_help;
+use crate::help_details::generate_help::show_generate_help;
+use crate::help_details::installers_help::{add_supported_installers, show_installers_help};
+use crate::help_details::now_help::show_now_help;
+use crate::help_details::sync_config_help::show_sync_config_help;
 use colored::Colorize;
 use std::fmt::Write;
-// Assuming this schema is defined elsewhere and contains the InstallerRegistry.
-// The use of 'crate::' indicates it's a module within the same project.
-use crate::schemas::help_schema::{InstallerRegistry, format_yaml_content};
 
 /// Dispatches the main help command based on the provided topic.
 ///
@@ -39,142 +41,6 @@ pub fn run(topic: Option<String>, detailed: bool, filter: Option<String>) {
         // If no topic is provided, display the default general help message.
         None => show_general_help(),
     }
-}
-
-/// Displays help for the `edit` command.
-///
-/// This function explains the purpose, usage, and options for the `edit` command,
-/// which is facilitates editing configuration and state file
-///
-/// # Arguments
-///
-/// * `detailed` - A `bool` flag to show comprehensive information about the command.
-fn show_edit_help(detailed: bool) {
-    let mut output = String::with_capacity(2048);
-
-    writeln!(output, "\n{}", "setup-devbox edit".bold().blue()).unwrap();
-    writeln!(output, "{}", "--------------------".bold().blue()).unwrap();
-    writeln!(output, "Edit setup-devbox configuration files.\n").unwrap();
-
-    // Add usage and options to the output string.
-    writeln!(output, "{}", "Usage:".bold().yellow()).unwrap();
-    writeln!(output, "  setup-devbox edit [OPTIONS]\n").unwrap();
-    writeln!(output, "{}", "Options:".bold().yellow()).unwrap();
-    writeln!(
-        output,
-        "  {}: Edit the state file. Use this option with caution.",
-        "--state".cyan()
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  {}: Edit a specific configuration file. [Supported Options: {}, {}, {} and {}]\n",
-        "--config <CONFIG_TYPE>".cyan(),
-        "tools".italic().cyan(),
-        "fonts".italic().cyan(),
-        "shell".italic().cyan(),
-        "settings".italic().cyan()
-    )
-    .unwrap();
-
-    // Conditionally add detailed or basic information based on the flag.
-    if detailed {
-        add_edit_detailed_info(&mut output);
-    } else {
-        add_edit_basic_examples(&mut output);
-    }
-
-    print!("{}", output);
-}
-
-/// Adds detailed information for the `edit` command to a mutable string.
-///
-/// This function provides a deeper dive into the `edit` command's functionality,
-/// explaining its role in state management and the files it uses.
-///
-/// # Arguments
-///
-/// * `output` - A mutable reference to the `String` where the content will be appended.
-fn add_edit_detailed_info(output: &mut String) {
-    writeln!(output, "{}", "Detailed Description:".bold().yellow()).unwrap();
-    writeln!(
-        output,
-        "  The '{}' command allows you to edit {} for your \
-    development environment.",
-        "edit".italic().cyan(),
-        "core configuration files".bold()
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  These files are used to manage tools, fonts, shell configurations, \
-    and operating system settings."
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  The command can also be used to edit the {}, though this should be \
-    done with caution\n",
-        "state file".bold()
-    )
-    .unwrap();
-    // Add behavior and examples information from other helper functions.
-    add_edit_behavior_info(output);
-    add_edit_basic_examples(output);
-}
-
-/// Adds behavior information for the `edit` command to a mutable string.
-///
-/// This function explains how the command manages editing the configuration
-/// and state files.
-///
-/// # Arguments
-///
-/// * `output` - A mutable reference to the `String` where the content will be appended.
-fn add_edit_behavior_info(output: &mut String) {
-    writeln!(output, "{}", "Behavior:".bold().yellow()).unwrap();
-    writeln!(
-        output,
-        "  • The command opens the specified file using the editor defined by your {} environment variable.",
-        "$EDITOR".italic().cyan()).unwrap();
-    writeln!(
-        output,
-        "  • It monitors the file's {} to detect any changes made during the editing session.",
-        "SHA".bold()
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  • If changes are detected, the {} command is automatically executed to apply them.",
-        "now".italic().cyan()
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  • If no changes are made to the file, the command exits without further action.\n"
-    )
-    .unwrap();
-}
-
-/// Adds detailed examples for the `edit` command to a mutable string.
-///
-/// # Arguments
-///
-/// * `output` - A mutable reference to the `String` where the content will be appended.
-fn add_edit_basic_examples(output: &mut String) {
-    writeln!(output, "{}", "Command Examples:".bold().yellow()).unwrap();
-    let examples = [
-        "setup-devbox edit --config tools",
-        "setup-devbox edit --config fonts",
-        "setup-devbox edit --config shell",
-        "setup-devbox edit --config settings",
-        "setup-devbox edit --state",
-    ];
-
-    for example in &examples {
-        writeln!(output, "  {}", example).unwrap();
-    }
-    writeln!(output).unwrap();
 }
 
 /// Displays an error message for an unknown help topic and lists available topics.
@@ -271,64 +137,6 @@ fn show_general_help() {
 
     // Print the fully built string to the console.
     print!("{}", output);
-}
-
-/// Adds the "Supported Installers" section to a mutable string.
-///
-/// This is a helper function that populates a `String` with a list of all
-/// supported installers, their descriptions, and any special notes.
-///
-/// # Arguments
-///
-/// * `output` - A mutable reference to the `String` where the content will be appended.
-fn add_supported_installers(output: &mut String) {
-    writeln!(output, "{}", "Supported Installers:".bold().yellow()).unwrap();
-
-    const INSTALLERS: [(&str, &str, &str); 10] = [
-        ("Brew", "Package manager for macOS/Linux (Homebrew)", ""),
-        ("Cargo", "Rust package manager for crates and binaries.", ""),
-        ("Fonts", "Nerd Fonts installer from GitHub releases.", ""),
-        ("Github", "Download tools from GitHub releases.", ""),
-        ("Go", "Go package installer using", "go install"),
-        ("Pip", "Python package installer using pip.", ""),
-        ("Rustup", "Rust toolchain installer and manager.", ""),
-        ("shell", "Shell configs", ".zshrc, .bashrc"),
-        ("URL", "Download and install from direct URLs.", ""),
-        (
-            "UV",
-            "Ultra-fast Python package installer",
-            "tool/pip/python",
-        ),
-    ];
-
-    // Find the longest installer name for padding
-    let max_width = INSTALLERS
-        .iter()
-        .map(|(name, _, _)| name.len())
-        .max()
-        .unwrap_or(0);
-
-    for (name, desc, extra) in &INSTALLERS {
-        write!(
-            output,
-            "  • {:width$} - {}",
-            name.bold().cyan(),
-            desc,
-            width = max_width
-        )
-        .unwrap();
-        if !extra.is_empty() {
-            if *name == "Go" {
-                write!(output, " '{}'.", extra.cyan()).unwrap();
-            } else if *name == "shell" {
-                write!(output, " ({} etc) and aliases.", extra.cyan()).unwrap();
-            } else if *name == "UV" {
-                write!(output, " ({} modes).", extra.cyan()).unwrap();
-            }
-        }
-        writeln!(output).unwrap();
-    }
-    writeln!(output).unwrap();
 }
 
 /// Adds the "Supported Environment Variables" section to a mutable string.
@@ -546,603 +354,6 @@ fn add_detailed_help_info(output: &mut String) {
     for example in &examples {
         writeln!(output, "  {}", example.italic()).unwrap();
     }
-}
-
-/// Displays help for supported installers with optional filtering and detail.
-///
-/// This function retrieves a list of all supported installers from the `InstallerRegistry`
-/// and formats the output. It respects the `--detailed` and `--filter` flags to
-/// provide tailored information to the user.
-///
-/// # Arguments
-///
-/// * `detailed` - A `bool` flag to show detailed information about each installer.
-/// * `filter` - An `Option<String>` to filter the list of installers by name.
-fn show_installers_help(detailed: bool, filter: Option<String>) {
-    println!("\n{}", "Supported Installers".bold().blue());
-    println!("{}", "--------------------".bold().blue());
-
-    // Print a general description based on the `detailed` flag.
-    if detailed {
-        println!(
-            "Comprehensive guide to all available installers with examples and configuration options.\n"
-        );
-    } else {
-        println!("Overview of available installers for setting up your development environment.\n");
-    }
-
-    // Retrieve all installers from the registry.
-    let installers = InstallerRegistry::get_all();
-    // Normalize the filter string to lowercase for case-insensitive matching.
-    let filter_str = filter.as_ref().map(|s| s.to_lowercase());
-    let mut matched_installers = Vec::new();
-
-    // Iterate through the installers and display them, applying the filter if it exists.
-    for installer in &installers {
-        if let Some(ref filter) = filter_str {
-            // Skip installers that don't match the filter.
-            if !installer.name.to_lowercase().contains(filter) {
-                continue;
-            }
-        }
-
-        // Call the `display` method on the installer object to print its details.
-        installer.display(detailed);
-        matched_installers.push(installer);
-    }
-
-    // If a filter was used but no installers matched, show available installers
-    if let Some(ref original_filter) = filter {
-        if matched_installers.is_empty() {
-            println!(
-                "{}: No installer found matching '{}'",
-                "Error".red(),
-                original_filter
-            );
-            println!("\n{}", "Available installers:".bold().yellow());
-
-            // Get all installers again to show the complete list
-            let all_installers = InstallerRegistry::get_all();
-
-            // Find the longest installer name for padding
-            let max_width = all_installers
-                .iter()
-                .map(|i| i.name.len())
-                .max()
-                .unwrap_or(0);
-
-            for installer in all_installers {
-                println!(
-                    "  • {:width$} - {}",
-                    installer.name.cyan(),
-                    installer.description,
-                    width = max_width
-                );
-            }
-        } else {
-            println!(
-                "\n{} Use without --filter to see all installers.",
-                "Tip:".yellow()
-            );
-        }
-    }
-}
-
-/// Displays help for the `now` command.
-///
-/// This function explains the purpose, usage, and options for the `now` command,
-/// which is responsible for installing and configuring the environment. It can
-/// show either a basic or a detailed view.
-///
-/// # Arguments
-///
-/// * `detailed` - A `bool` flag to show comprehensive information about the command.
-fn show_now_help(detailed: bool) {
-    let mut output = String::with_capacity(2048);
-
-    writeln!(output, "\n{}", "setup-devbox now".bold().blue()).unwrap();
-    writeln!(output, "{}", "--------------------".bold().blue()).unwrap();
-    writeln!(
-        output,
-        "Install and configure your development environment based on configuration files.\n"
-    )
-    .unwrap();
-
-    // Add usage and options to the output string.
-    writeln!(output, "{}", "Usage:".bold().yellow()).unwrap();
-    writeln!(output, "  setup-devbox now [OPTIONS]\n").unwrap();
-
-    writeln!(output, "{}", "Options:".bold().yellow()).unwrap();
-    writeln!(
-        output,
-        "  {} Use custom configuration file (default: {})",
-        "--config <PATH>".cyan(),
-        "~/.setup-devbox/config.yaml".cyan()
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  {} Use custom state file (default: {})",
-        "--state <PATH>".cyan(),
-        "~/.setup-devbox/state.json".cyan()
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  {} Force update all tools marked as '{}' version. Overrides the configuration \"update_latest_only_after\"\n",
-        "--update-latest".cyan(),
-        "@latest".cyan()
-    )
-        .unwrap();
-
-    // Conditionally add detailed or basic information based on the flag.
-    if detailed {
-        add_now_detailed_info(&mut output);
-    } else {
-        add_now_basic_examples(&mut output);
-    }
-
-    print!("{}", output);
-}
-
-/// Adds detailed information for the `now` command to a mutable string.
-///
-/// This function provides a deeper dive into the `now` command's functionality,
-/// explaining its role in state management and the files it uses.
-///
-/// # Arguments
-///
-/// * `output` - A mutable reference to the `String` where the content will be appended.
-fn add_now_detailed_info(output: &mut String) {
-    writeln!(output, "{}", "Detailed Description:".bold().yellow()).unwrap();
-    writeln!(
-        output,
-        "  The 'now' command is the main installation command that reads your configuration"
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  files and installs all specified tools, fonts, and configurations. It maintains"
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  a state file to track what's been installed and their versions.\n"
-    )
-    .unwrap();
-
-    // Explain the configuration file and its components.
-    writeln!(output, "{}", "Configuration File:".bold().yellow()).unwrap();
-    writeln!(
-        output,
-        "The configuration file ({}) is collection other configuration files paths:",
-        "config.yaml".cyan()
-    )
-    .unwrap();
-    let config_items = [
-        (
-            "tools.yaml",
-            "Tools and their versions needs to be installed",
-        ),
-        ("fonts.yaml", "Fonts to install"),
-        ("shellrc.yaml", "Shell configurations to be applied"),
-        ("settings.yaml", "OS-specific settings changes to be made"),
-    ];
-
-    for (file, desc) in &config_items {
-        writeln!(output, "  • {} ({})", desc, file.cyan()).unwrap();
-    }
-    writeln!(output).unwrap();
-
-    // Explain the state management file.
-    writeln!(output, "{}", "State Management:".bold().yellow()).unwrap();
-    writeln!(output, "The state file ({}) tracks:", "state.yaml".cyan()).unwrap();
-    let state_items = [
-        "Currently installed tools and versions",
-        "Installation timestamps",
-        "Success/failure status of installations",
-    ];
-
-    for item in &state_items {
-        writeln!(output, "  • {}", item).unwrap();
-    }
-    writeln!(output).unwrap();
-
-    // Add behavior and examples information from other helper functions.
-    add_now_behavior_info(output);
-    add_now_detailed_examples(output);
-}
-
-/// Adds detailed examples for the `now` command to a mutable string.
-///
-/// # Arguments
-///
-/// * `output` - A mutable reference to the `String` where the content will be appended.
-fn add_now_detailed_examples(output: &mut String) {
-    writeln!(output, "\n{}", "Command Examples:".bold().yellow()).unwrap();
-    let examples = [
-        "setup-devbox now",
-        "setup-devbox now --config ./my-config.yaml",
-        "setup-devbox now --update-latest",
-        "setup-devbox now --config custom.yaml --state custom-state.json",
-    ];
-
-    for example in &examples {
-        writeln!(output, "  {}", example).unwrap();
-    }
-    writeln!(output).unwrap();
-
-    add_configuration_examples(output);
-}
-
-/// Adds configuration file examples to a mutable string.
-///
-/// This function provides detailed examples of what the configuration files look like,
-/// showing the structure and format users should follow.
-///
-/// # Arguments
-///
-/// * `output` - A mutable reference to the `String` where the content will be appended.
-///
-fn add_configuration_examples(output: &mut String) {
-    writeln!(output, "{}", "Configuration Examples:".bold().yellow()).unwrap();
-    writeln!(
-        output,
-        "Here are examples of the configuration files structure:\n"
-    )
-    .unwrap();
-
-    // Main config.yaml example
-    writeln!(output, "{}:\n", "config.yaml".bold().cyan()).unwrap();
-    writeln!(output, "{}", "```yaml".green()).unwrap();
-    let config_yaml = r#"# Tells SDB where to find the tools configuration
-tools: /Users/<<user>>/.config/setup-devbox/configs/tools.yaml
-# Tells SDB where to find the settings configuration
-settings: /Users/<<user>>/.config/setup-devbox/configs/settings.yaml
-# Tells SDB where to find the shell configuration
-shellrc: /Users/<<user>>/.config/setup-devbox/configs/shellrc.yaml
-# Tells SDB where to find the fonts configuration
-fonts: /Users/<<user>>/.config/setup-devbox/configs/fonts.yaml"#;
-    write!(output, "{}", format_yaml_content(config_yaml)).unwrap();
-    writeln!(output, "{}\n", "```".green()).unwrap();
-
-    // tools.yaml example
-    writeln!(output, "{}:\n", "tools.yaml".bold().cyan()).unwrap();
-    writeln!(output, "{}", "```yaml".green()).unwrap();
-    let tools_yaml = r#"update_latest_only_after: "7 days"
-tools:
-
-  # This will install rustup
-  - name: rustup
-    source: brew # Install rustup via Homebrew
-
-  # Example: Install pyenv
-  - name: pyenv
-    source: brew
-    options:
-      - --head"#;
-    write!(output, "{}", format_yaml_content(tools_yaml)).unwrap();
-    writeln!(output, "{}\n", "```".green()).unwrap();
-
-    // fonts.yaml example
-    writeln!(output, "{}:\n", "fonts.yaml".bold().cyan()).unwrap();
-    writeln!(output, "{}", "```yaml".green()).unwrap();
-    let fonts_yaml = r#"fonts:
-
-  # name: name of the font
-  # version: specific version of the font
-  # source: from where to download the fonts, (GitHub only)
-  # repo: repository name
-  # tag: the specific release tag for this font version
-  # install_only: only install the one mentioned, default is `all`
-  - name: 0xProto
-    version: "3.4.0"
-    source: github
-    repo: ryanoasis/nerd-fonts
-    tag: v3.4.0
-    install_only: ['regular', 'Mono']"#;
-    write!(output, "{}", format_yaml_content(fonts_yaml)).unwrap();
-    writeln!(output, "{}\n", "```".green()).unwrap();
-
-    // shellrc.yaml example
-    writeln!(output, "{}:\n", "shellrc.yaml".bold().cyan()).unwrap();
-    writeln!(output, "{}", "```yaml".green()).unwrap();
-    let shellrc_yaml = r#"run_commands:
-  # Supported Shells: `zsh` and `bash`
-  shell: "zsh" # or "bash"
-  run_commands:
-    # Exports Section - Environment variables
-    # Any environment variables which needs to be exported
-    - command: |
-        export EDITOR="zed"
-        export VISUAL="zed"
-      section: Exports
-
-    # Paths Section - PATH modifications
-    - command: export PATH="$HOME/bin:$PATH"
-      section: Paths
-
-    # Evals Section - Command evaluations
-    - command: eval "$(pyenv init - zsh)"
-      section: Evals
-  aliases:
-    # Alias Section - For all aliases
-    - name: sd
-      value: setup-devbox"#;
-    write!(output, "{}", format_yaml_content(shellrc_yaml)).unwrap();
-    writeln!(output, "{}\n", "```".green()).unwrap();
-
-    // settings.yaml example
-    writeln!(output, "{}:\n", "settings.yaml".bold().cyan()).unwrap();
-    writeln!(output, "{}", "```yaml".green()).unwrap();
-    let settings_yaml = r#"settings:
-  macos: # Settings specifically for macOS
-    # A common domain for global macOS settings
-    - domain: NSGlobalDomain
-      # The specific setting key: show file extensions
-      key: AppleShowAllExtensions
-      # Set its value to true
-      value: "true"
-      # This setting expects a boolean value
-      type: bool"#;
-    write!(output, "{}", format_yaml_content(settings_yaml)).unwrap();
-    writeln!(output, "{}\n", "```".green()).unwrap();
-}
-
-/// Adds behavior information for the `now` command to a mutable string.
-///
-/// This function explains how the command manages installations, including
-/// skipping, updating, and state file backups.
-///
-/// # Arguments
-///
-/// * `output` - A mutable reference to the `String` where the content will be appended.
-fn add_now_behavior_info(output: &mut String) {
-    writeln!(output, "{}", "Behavior:".bold().yellow()).unwrap();
-    writeln!(
-        output,
-        "  • Skips already installed tools unless versions differ"
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  • Updates tools marked with '{}' only after configured time period",
-        "latest".cyan()
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  • Use {} to force update all tools '{}' version immediately",
-        "--update-latest".italic().cyan(),
-        "@latest".cyan()
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  • Creates backup of state file before major changes"
-    )
-    .unwrap();
-}
-
-/// Adds basic examples for the `now` command to a mutable string.
-///
-/// This is a simple version of the examples, shown when `detailed` is false.
-///
-/// # Arguments
-///
-/// * `output` - A mutable reference to the `String` where the content will be appended.
-fn add_now_basic_examples(output: &mut String) {
-    writeln!(output, "{}", "Examples:".bold().yellow()).unwrap();
-    writeln!(
-        output,
-        "  setup-devbox now                    # Use default config"
-    )
-    .unwrap();
-    writeln!(
-        output,
-        "  setup-devbox now --update-latest    # Force update latest versions"
-    )
-    .unwrap();
-}
-
-/// Displays help for the `generate` command.
-///
-/// This function explains the purpose, usage, and options for the `generate` command,
-/// which creates starter configuration files.
-///
-/// # Arguments
-///
-/// * `detailed` - A `bool` flag to show comprehensive information.
-fn show_generate_help(detailed: bool) {
-    println!("{}", "setup-devbox generate".bold().blue());
-    println!("Generate default configuration and state files to get started.\n");
-
-    println!("{}", "Usage:".bold());
-    println!("  setup-devbox generate [OPTIONS]\n");
-
-    println!("{}", "Options:".bold());
-    println!(
-        "  --config <PATH>  Save configuration to custom path (default: ~/.setup-devbox/config.yaml)"
-    );
-    println!(
-        "  --state <PATH>   Save state file to custom path (default: ~/.setup-devbox/state.json)\n"
-    );
-
-    if detailed {
-        show_generate_detailed_help();
-    } else {
-        show_generate_basic_examples();
-    }
-}
-
-/// Shows detailed help for the `generate` command.
-///
-/// The `generate` command creates template configuration files that you can
-/// customize for your development environment. It's the recommended way to
-/// get started with setup-devbox.
-///
-/// **Generated Files:**
-///   - `config.yaml`
-///   - `tools.yaml`
-///   - `fonts.yaml`
-///   - `settings.yaml`
-///   - `shellrc.yaml`
-///   - `state.json`
-///
-fn show_generate_detailed_help() {
-    println!("{}", "Detailed Description:".bold());
-    println!("The 'generate' command creates template configuration files that you can");
-    println!("customize for your development environment. It's the recommended way to");
-    println!("get started with setup-devbox.\n");
-
-    // List the files that are created.
-    println!("{}", "Generated Files:".bold());
-    println!(
-        "  • config.yaml    - Main configuration with references to other configuration files"
-    );
-    println!("  • tools.yaml     - Tools configuration with example tools");
-    println!("  • fonts.yaml     - Fonts configuration with example fonts");
-    println!("  • shellrc.yaml   - Shell configuration with example shell configuration");
-    println!(
-        "  • settings.yaml  - OS Setting configuration with example configuration (Supported OS: macOS"
-    );
-    println!("  • state.json     - Empty state file to track installations\n");
-
-    // Explain the contents of the generated configuration.
-    println!("{}", "Configuration Structure:".bold());
-    println!("The generated config includes examples for:");
-    let config_examples = [
-        "Common development tools (git, docker, node, etc.)",
-        "Popular programming languages and their toolchains",
-        "Essential fonts for development",
-        "Shell configurations and aliases",
-        "OS-specific tweaks and settings",
-    ];
-
-    for example in &config_examples {
-        println!("  • {}", example);
-    }
-    println!();
-
-    // Provide usage examples.
-    println!("{}", "Examples:".bold());
-    println!("  setup-devbox generate");
-    println!("  setup-devbox generate --config ./project-config.yaml");
-    println!("  setup-devbox generate --config dev.yaml --state dev-state.json\n");
-
-    // Outline the typical workflow for a new user.
-    println!("{}", "Next Steps:".bold());
-    println!("  1. Run 'setup-devbox generate' to create default configs");
-    println!("  2. Edit the generated config.yaml to match your needs");
-    println!("  3. Run 'setup-devbox now' to install everything");
-}
-
-/// Shows basic examples for the `generate` command.
-fn show_generate_basic_examples() {
-    println!("{}", "Examples:".bold());
-    println!("  setup-devbox generate                    # Create default configs");
-    println!("  setup-devbox generate --config custom.yaml # Custom location");
-}
-
-/// Displays help for the `sync-config` command.
-///
-/// This function explains the purpose, usage, and options for the `sync-config` command,
-/// which reconstructs configuration files from a state file.
-///
-/// # Arguments
-///
-/// * `detailed` - A `bool` flag to show comprehensive information.
-fn show_sync_config_help(detailed: bool) {
-    println!("{}", "setup-devbox sync-config".bold().blue());
-    println!("Synchronize or generate configurations from an existing state file.\n");
-
-    println!("{}", "Usage:".bold());
-    println!("  setup-devbox sync-config [OPTIONS]\n");
-
-    println!("{}", "Options:".bold());
-    println!("  --state <PATH>      Path to state file (default: ~/.setup-devbox/state.json)");
-    println!(
-        "  --output-dir <PATH> Output directory for configs (default: ~/.setup-devbox/configs)\n"
-    );
-
-    if detailed {
-        show_sync_config_detailed_help();
-    } else {
-        show_sync_config_basic_examples();
-    }
-}
-
-/// Shows detailed help for the `sync-config` command.
-///
-/// The 'sync-config' command helps you recreate configuration files from an existing state file.
-/// This is useful for:
-///   1. Recovering lost configuration files
-///   2. Sharing your setup across multiple machines (state.json)
-///   3. Creating configuration templates from working setups
-///
-fn show_sync_config_detailed_help() {
-    println!("{}", "Detailed Description:".bold());
-    println!("The 'sync-config' command helps you recreate configuration files from");
-    println!("an existing state file. This is useful for:");
-    let use_cases = [
-        "Recovering lost configuration files",
-        "Sharing your setup across multiple machines",
-        "Creating configuration templates from working setups",
-    ];
-
-    for case in &use_cases {
-        println!("  • {}", case);
-    }
-    println!();
-
-    // Provide additional use cases.
-    println!("{}", "Use Cases:".bold());
-    let additional_cases = [
-        "Backup and restore configurations",
-        "Team configuration sharing",
-        "Migration between development environments",
-        "Creating standardized setups",
-    ];
-
-    for case in &additional_cases {
-        println!("  • {}", case);
-    }
-    println!();
-
-    // Describe the generated output.
-    println!("{}", "Generated Output:".bold());
-    println!("Creates configuration files based on your state file:");
-    let outputs = [
-        "Reconstructed config.yaml with current tool versions",
-        "Shell configuration files (.zshrc, .bashrc)",
-        "Tool-specific configuration files",
-        "Documentation of your current setup",
-    ];
-
-    for output in &outputs {
-        println!("  • {}", output);
-    }
-    println!();
-
-    // Provide usage examples.
-    println!("{}", "Examples:".bold());
-    println!("  setup-devbox sync-config");
-    println!("  setup-devbox sync-config --state ./backup-state.json");
-    println!("  setup-devbox sync-config --output-dir ./team-configs");
-    println!("  setup-devbox sync-config --state prod.json --output-dir ./prod-configs\n");
-
-    // Explain the typical workflow.
-    println!("{}", "Workflow:".bold());
-    println!("  1. Run 'setup-devbox now' to install and track your tools");
-    println!("  2. Use 'setup-devbox sync-config' to generate shareable configs");
-    println!("  3. Share the generated configs with your team or other machines");
-}
-
-/// Shows basic examples for the `sync-config` command.
-fn show_sync_config_basic_examples() {
-    println!("{}", "Examples:".bold());
-    println!("  setup-devbox sync-config                 # Use default paths");
-    println!("  setup-devbox sync-config --state backup.json # From backup");
 }
 
 /// Displays help for the `version` command.
