+++ src/help_details/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod edit_help;
+pub(crate) mod generate_help;
+pub(crate) mod installers_help;
+pub(crate) mod now_help;
+pub(crate) mod sync_config_help;
